Add HelpCommand to TagsPlugin and update tag deletion logic
In this update, I have added a new functionality into TagsPlugin by adding the HelpCommand. This feature will help to display bot documentation effectively. In addition, we've made a modification to the tag deletion logic so that the author of a tag also has the ability to delete the tag. Besides this, new '.dockerignore' file was created, and the user documentation was updated to reflect these changes.

Signed-off-by: zueripat <ruedepatrick@gmail.com>